Fix two problems with int32 <-> int32_t conversion when compiled in -pedantic mode inside Native Client. BUG=none TEST=none
Review URL: http://codereview.chromium.org/2834023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@50760 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 42af38a25b076918102de7d21c0876015f8d01be

+++ gpu/pgl/command_buffer_pepper.cc
@@ -131,11 +131,11 @@ void CommandBufferPepper::SetGetOffset(int32 get_offset) {
 }
 
 int32 CommandBufferPepper::CreateTransferBuffer(size_t size) {
-  int32 id;
+  int32_t id;
   if (NPERR_NO_ERROR != device_->createBuffer(npp_, context_, size, &id))
     return -1;
 
-  return id;
+  return static_cast<int32>(id);
 }
 
 void CommandBufferPepper::DestroyTransferBuffer(int32 id) {
@@ -165,7 +165,7 @@ void CommandBufferPepper::SetParseError(
 }
 
 gpu::error::Error CommandBufferPepper::GetCachedError() {
-  int32 attrib_list[] = {
+  int32_t attrib_list[] = {
     NPAttrib_Error, 0,
     NPAttrib_End
   };
